pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nayak"
version = "0.2.0"
description = "Open source runtime engine for autonomous robots and AI agents that work in the physical world."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "NAYAK Authors" }]
keywords = ["autonomous", "agents", "robotics", "ai", "browser-automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # AI providers
    "openai>=1.50.0",          # NVIDIA NIM uses OpenAI-compatible API
    "httpx>=0.27.0",           # Ollama local uses httpx directly
    # Browser automation
    "playwright>=1.49.0",
    # CLI & output
    "typer>=0.15.0",
    "rich>=13.9.0",
    # Memory
    "aiosqlite>=0.20.0",
    # Config
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
nayak = "nayak.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["nayak"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
